requirements.txt
rich
playwright
requests
pyfiglet
setuptools
dnspython
markdown-it-py